third_party/WebKit/LayoutTests/http/tests/inspector-protocol/request-mixed-content-status-blockable-expected.txt

CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/inspector-protocol/resources/active-mixed-content-iframe.html' was loaded over HTTPS, but requested an insecure script 'http://127.0.0.1:8000/inspector-protocol/resources/blank.js'. This content should also be served over HTTPS.
Tests that willSendRequest contains the correct mixed content status for active mixed content.

Network agent enabled
Mixed content type of https://127.0.0.1:8443/inspector-protocol/resources/active-mixed-content-iframe.html: none
Mixed content type of http://127.0.0.1:8000/inspector-protocol/resources/blank.js: blockable
